Final review

+++ PlantFriend.ui/src/components/Cards/PlantCard.js
@@ -19,10 +19,11 @@ import LogCard from './LogCard';
 function PlantCard({
   setPlants,
   setUserPlants,
+  userPlantIds,
   ...rest
 }) {
-  const [modalStatus, setModalStatus] = useState(false);
-  const modalToggle = () => setModalStatus(!modalStatus);
+  const [modalStatus0, setModalStatus0] = useState(false);
+  const modalToggle0 = () => setModalStatus0(!modalStatus0);
 
   const [modalStatus2, setModalStatus2] = useState(false);
   const modalToggle2 = () => setModalStatus2(!modalStatus2);
@@ -37,12 +38,14 @@ function PlantCard({
   const toggleLogContainer = () => setLogContainerStatus(!logContainerStatus);
 
   const [plantLogs, setPlantLogs] = useState([]);
-  useEffect(() => {
-    getLogsByUserPlantId(rest.userPlantId).then(setPlantLogs);
-  }, []);
 
   const [desc, setDesc] = useState(false);
   const toggleDesc = () => setDesc(!desc);
+
+  const userPageCheck = (useLocation().pathname === '/user');
+  useEffect(() => {
+    if (userPageCheck) getLogsByUserPlantId(rest.userPlantId).then(setPlantLogs);
+  }, []);
 
   return (
     <Col className="col-sm-4">
@@ -79,7 +82,7 @@ function PlantCard({
                 <Button onClick={() => modalToggle4()}><i className='fas fa-plus-circle'></i></Button>
                 <Button onClick={() => {
                   // eslint-disable-next-line
-                  const result = window.confirm('Are you sure you want to remove your plant? Its journal and all of its information will be permanently deleted.');
+                  const result = window.confirm('Are you sure you want to remove your plant? Its journal and all of its information will be permanently removed.');
                   if (result) deleteUserPlant(rest.userPlantId, rest.userId).then(setUserPlants);
                 }}><i className="fas fa-trash-alt"></i></Button>
               </Row>
@@ -87,18 +90,23 @@ function PlantCard({
           {useLocation().pathname === '/plants/'
             && <>
               <Row className='justify-content-around'>
-                <Button onClick={() => modalToggle()}><i className="fas fa-edit"></i></Button>
-                <Button onClick={() => modalToggle2()}><i className="fas fa-plus-square"></i></Button>
+                <Button onClick={() => modalToggle0()}><i className="fas fa-edit"></i></Button>
+                <Button onClick={() => modalToggle2()}><i className="fas fa-plus-circle"></i></Button>
                 <Button onClick={() => {
-                  // eslint-disable-next-line
-                  const result = window.confirm('Are you sure? All of your research on this plant will be permanently deleted, and you&aps;ll no longer be able to add it to your profile.');
-                  if (result) deletePlant(rest.id).then(setPlants).then(toggleLogContainer);
+                  if (userPlantIds.includes(rest.id)) {
+                    // eslint-disable-next-line
+                    window.alert(`You cannot remove a plant that you currently own. Please remove any ${rest.name} from your profile first.`);
+                  } else {
+                    // eslint-disable-next-line
+                    const result = window.confirm(`Are you sure? All of your research on this plant (${rest.name}) will be permanently removed, and you'll no longer be able to add it to your profile.`);
+                    if (result) deletePlant(rest.id).then(setPlants).then(toggleLogContainer);
+                  }
                 }}><i className="fas fa-trash-alt"></i></Button>
               </Row>
             </>}
         </Col>
       </Card>
-      <Collapse isOpen={logContainerStatus}>
+      <Collapse isOpen={logContainerStatus && useLocation().pathname === '/user'}>
         <div className='plant-color'>Journal:
           {plantLogs.length === 0
             && <div>You currently have no journal entries.</div>
@@ -113,11 +121,14 @@ function PlantCard({
               entry={log.entry}
               entryDate={log.entryDate}
               setPlantLogs={setPlantLogs}
+              petName={rest.petName}
+              name={rest.name}
             />
           ))}
         </div>
       </Collapse>
       <FormModal
+        key={rest.id + rest.name}
         id={rest.id}
         name={rest.name}
         water={rest.water}
@@ -129,14 +140,16 @@ function PlantCard({
         imageUrl={rest.imageUrl}
         careNeeds={rest.careNeeds}
         light={rest.light}
-        modalStatus={modalStatus} modalToggle={modalToggle} modalTitle='Update Plant Research' setPlants={setPlants}
+        modalStatus={modalStatus0} modalToggle={modalToggle0} modalTitle='Update Plant Research' setPlants={setPlants}
       />
       <FormModal
+        key={rest.id + rest.userId}
         plantId={rest.id}
         userId={rest.userId}
         modalStatus={modalStatus2} modalToggle={modalToggle2} modalTitle='Add Plant to Profile' setUserPlants={setUserPlants}
       />
       <FormModal
+        key={rest.userPlantId + rest.id}
         id={rest.userPlantId}
         plantId={rest.id}
         userId={rest.userId}
@@ -147,6 +160,7 @@ function PlantCard({
         modalStatus={modalStatus3} modalToggle={modalToggle3} modalTitle='Update Your Plant' setUserPlants={setUserPlants}
       />
       <FormModal
+        key={rest.userPlantId}
         userPlantId={rest.userPlantId}
         modalStatus={modalStatus4} modalToggle={modalToggle4} modalTitle='Add A New Journal Entry' setPlantLogs={setPlantLogs}
       />
@@ -177,4 +191,5 @@ PlantCard.propTypes = {
   dateCreated: PropTypes.string,
   initialAgeDays: PropTypes.number,
   ageStage: PropTypes.string,
+  userPlantIds: PropTypes.array
 };
